+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGELOG
 
+## dev
+- removed multi primary/secondary management
+- added 'editable:middlemarker:onmousedown' event
+
 ## 0.1.0
 - initial release